Refactored the RESTfulModelFormMixin
The RESTfulModelFormMixin class is now just the ModelForm class. It
simply inherits from the Django ModelForm class and provides a bit of
code to set the default values from a Model object on the incoming
data object allowing validation to work for RESTful requests where not
all data is sent everytime.