Updated the RESTfulResponse class to allow HttpResponse objects to be returned
This allows Resources that return regular HttpResponse objects to
continue to operate as if content negotiation was not being used
rather than causing an exception to be thrown. This should make it
much easier to use, and tranition to using, content negotiation since
full resources can decorated and all methods will still naturally
work.