Updated RESTfulResponse to set all data to the variable 'context' in a template
If a template is used during content negotiation for transforming data
into the desired representation, a variable named 'context' that holds
all of the data returned by the resource is available from within the
template. Also, if no data is returned, rendering the template is
skipped. This matches the current behavior when a callable is used in
place of a template. Currently the callable is only called when data
is available.