request objects now contain QueryDicts that correspond to the HTTP method used
Since it is technically possible to have an entity body for any HTTP
request according to the HTTP spec, the Resource class will populate a
QueryDict and add it to the Request object under the name of the HTTP
method that the request was determined to use (this could be different
from the actual HTTP method depending on whether or not the _method
attribute was sent in the request). Also, the POST QueryDict is now
unmutable even when the _method attribute is sent. Previously this was
untrue, since a copy was made of the POST QueryDict in order to 'pop'
the _method attribute off of the QueryDict. Finally, he QueryDict for
the HTTP method used is a pointer to the POST QueryDict instead of a
completely different copy of it. This should be fine since the
QueryDict is now unmutable even when the _method attribute is sent.