Removed the exceptions middleware
The main reason for this is to get the HttpError exception working
with content negotiation. Since content negotation is taken care of by
a decorator on the view function, and the exception middleware doesn't
have access to the decorator, handling an HttpError had to be moved
into the RESTfulResponse class and the dispatch method of the
Resource. I'd still prefer to handle this in middleware instead of
having the handling spread out over two classes, but right now I can't
think of a good way to do it.