Fixed a bug with the ModelForm implementation
When sending a GET request that uses the _method option to make the
request act as a PUT, the request.PUT QueryDict object is not
populated. To deal with this, it is suggested that the user use
request.REQUEST to make sure that they can access the data. The
problem with doing this though is that the request.REQUEST object is
not a QueryDict, but a MergeDict and the MergeDict object does not
support updates. The solution was to convert the data object in the
ModelForm to a simple dict. This shouldn't cause any issues since the
BaseForm object in the Django source actually sets the data object to
either the given data object or an empty dict.